重构 DPI 计算逻辑，删除 MultiAtlasGenerator 类
在 `Window.cs` 文件中，修改了 DPI 的计算逻辑，增加了对 DPI 为 0 的处理，避免除以 0 的错误。

在 `MainView.cs` 文件中：
  * 删除了 `_multiAtlasGenerator` 字段。
  * 修改了 `_layout` 的初始化方式，从使用 `MultiAtlasGenerator` 生成改为从 JSON 文件解析。
  * 删除了与 `_multiAtlasGenerator` 相关的代码，简化了纹理坐标的计算逻辑。

在 `CharacterRange.cs` 文件中，删除了 `CharacterRange` 结构体及其相关代码。

在 `Metrics.cs` 文件中，将 `EmSize` 属性的类型从 `int` 改为 `float`。

在 `MultiAtlasGenerator.cs` 文件中，删除了整个 `MultiAtlasGenerator` 类及其相关代码。

在 `launchSettings.json` 文件中，新增了启动配置文件，包含 `Tests.MultiViewports` 和 `WSL` 两个配置。

+++ Source/Tests/Tests.SDFFontTexture/Models/Metrics.cs
@@ -2,7 +2,7 @@
 
 internal sealed class Metrics
 {
-    public int EmSize { get; set; }
+    public float EmSize { get; set; }
 
     public float LineHeight { get; set; }
